moved circom code to new directory and removed it from primitives
