+++ primitives/src/verifying/circom.rs
@@ -28,26 +28,6 @@ use once_cell::sync::OnceCell;
 use arkworks_native_gadgets::to_field_elements;
 // use sp_core::U256;
 
-/// Verifies a given RLN proof
-///
-/// # Errors
-///
-/// Returns a [`ProofError`] if verifying fails. Verification failure does not
-/// necessarily mean the proof is incorrect.
-pub fn verify_proof(
-	verifying_key: &VerifyingKey<Bn254>,
-	proof: &ArkProof<Bn254>,
-	inputs: Vec<Fr>,
-) -> Result<bool, ProofError> {
-	// Check that the proof is valid
-	let pvk = prepare_verifying_key(verifying_key);
-	//let pr: ArkProof<Curve> = (*proof).into();
-
-	let verified = ark_verify_proof(&pvk, proof, &inputs)?;
-
-	Ok(verified)
-}
-
 pub struct CircomVerifierBn254;
 
 #[derive(Debug)]
